pyproject.toml
[project]
name = "discord-gm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.13",
    "discord-py>=2.5.2",
    "openai==0.28.1",
    "python-dotenv>=1.0.1",
]
